thumbnail の表示エラー解決 (#22)
* fix: thumbnail の表示エラー解決

* test: Add CachedNetworkImage mock and update widget tests for setlist display

* fix: Enable useOldImageOnUrlChange for CachedNetworkImage in setlist page

* test: errorを表示

* chore: Image.networkに変更

* fix: Improve image loading handling in setlist page by adding loading and error indicators

* chore: remove cached_network_image
